完成websocket代码，rabbitmq基本配置与接入

+++ internal/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	ID       int    `json:"id" gorm:"primaryKey"` 
+	ID       uint64    `json:"id" gorm:"primaryKey"` 
 	Username string `json:"username" gorm:"unique;not null"`
 	Password string `json:"password" gorm:"not null"`
 }
